+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * Add bash completion for `logs` command
 * Format status output to make clearer
 * Add new Nomad verison 1.1.3
+* Return error message when resource has no name
 
 ## version v0.3.1